move new url stuff to urlutils file
