--- app/src/utils/useSearchParams.ts
@@ -1,10 +1,0 @@
-import { useLocation } from "react-router-dom";
-
-function useSearchParams() {
-  // Note: URLSearchParams doesn't support IE, if we keep this code and we want
-  // to support IE we'll need to change it.
-  const params = new URLSearchParams(useLocation().search);
-  return Object.fromEntries(params.entries());
-}
-
-export default useSearchParams;
